LPS-119571 rename TranslationManagerWrapper to TranslationManagerEntry
