+++ modules/apps/app-builder/app-builder-web/src/main/resources/META-INF/resources/js/pages/entry/TranslationManagerEntry.es.js
